requirements.txt
google-genai>=1
pydantic
pdfminer.six
Pillow
pytesseract
